Geolocalizacion por network

+++ app/src/main/java/com/tallerii/match/core/UserProfile.java
@@ -20,8 +20,8 @@ public class UserProfile implements Serializable{
     private String photo = "";
     private int age = 18;
 
-    private int latitude;
-    private int longitude;
+    private double latitude;
+    private double longitude;
 
     private String id = "";
 
@@ -115,19 +115,19 @@ public class UserProfile implements Serializable{
         this.photo = photo;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
